move serializer fields into fields.py
